Вынесет утилитарные функции для задач в отдельный модуль
