+++ src/utils/common.js
@@ -1,5 +1,3 @@
-import dayjs from 'dayjs';
-
 export const Key = {
   ESC: 'Esc',
   ESCAPE: 'Escape',
@@ -10,12 +8,6 @@ export const Evt = {
   KEYDOWN: 'keydown',
   SUBMIT: 'submit',
   CHANGE: 'change',
-};
-
-export const SortType = {
-  DAY: 'sort-day',
-  TIME: 'sort-time',
-  PRICE: 'sort-price',
 };
 
 export const getRandomInteger = (a = 0, b = 1) => {
@@ -43,14 +35,4 @@ export const updateItem = (items, updatedItem) => {
     updatedItem,
     ...items.slice(index + 1),
   ];
-};
-
-export const sortPointPrice = (pointA, pointB) => pointB.base_price - pointA.base_price;
-
-export const sortPointDay = (pointA, pointB) => dayjs(pointA.date_from).diff(dayjs(pointB.date_from));
-
-export const sortPointTime = (pointA, pointB) => {
-  const durationA = pointA.date_to - pointA.date_from;
-  const durationB = pointB.date_to - pointB.date_from;
-  return durationB - durationA;
 };
